inquirer.js: Task deletition confirmation message added

+++ app.js
@@ -1,6 +1,6 @@
 require('colors');
 
-const {inquirerMenu, pause, readInput, listDeletingTask} = require('./helpers/inquirer');
+const {inquirerMenu, pause, readInput, listDeletingTask, confirm} = require('./helpers/inquirer');
 const {saveInDB, readDB} = require('./helpers/saveInfo');
 const Task = require('./models/task');
 const Tasks = require('./models/tasks');
@@ -38,7 +38,12 @@ const main = async() => {
                 break;
             case '6':
                 const id = await listDeletingTask(tasks.listArray);
-                tasks.deleteTasks(id);
+                const isOk = await confirm('Are you sure?');
+                
+                if(isOk){
+                    tasks.deleteTasks(id);
+                }
+                
                 break;
         }
